fix: change default dtype from float16 to bfloat16
Some models require bfloat16 for proper execution. Using float16 as default could lead to compatibility issues or performance degradation for those models. Changing the default to bfloat16 ensures broader model support.

+++ minivllm/engine/inference_executor.py
@@ -122,7 +122,7 @@ class InferenceExecutor:
                     'float32': torch.float32,
                     'bfloat16': torch.bfloat16,
                 }
-                dtype = dtype_map.get(dtype, torch.float16)
+                dtype = dtype_map.get(dtype, torch.bfloat16)
 
             hf_config = self.config.hf_config
             head_dim = getattr(
